+++ src/lib/components/json-schema/index.ts
@@ -3,7 +3,8 @@ export * from "./AttachSchemaCard";
 export * from "./EditSchemaButtons";
 export * from "./JsonSchemaModal";
 export * from "./UploadTemplate";
-export * from "./ViewSchemaButton";
+export * from "./view/ViewSchemaModal";
 export * from "./MessageInputSwitch";
 export * from "./MessageInputContent";
 export * from "./section/SchemaInputSection";
+export * from "./upload";
