+++ src/lib/components/json-schema/upload/index.ts
@@ -1,0 +1,2 @@
+export * from "./UploadSchema";
+export * from "./UploadSchemaSection";
